Remove SSDP; fix mDNS hostname after HomeKit; clang-tidy cleanups
Remove the SSDP module (~950 LOC + AsyncUDP dep). It was unused — HomeKit
advertises via _hap._tcp, the web UI is reached by IP/hostname, and no
client of this firmware depends on UPnP discovery. Drops main/SSDP.cpp/h,
the /upnp.xml route, the ssdpBroadcast setting, and the SSDP_* defines.
shades.cfg bumped to v25 with backward-compat read for v24; existing
devices shed the legacy NVS key on next boot. AppConfig.h was reaching
ControllerNetwork.cpp and web/Web.cpp transitively via SSDP.h — replaced
with direct includes.

Fix <hostname>.local resolution. esp-homekit-sdk's hap_mdns_init()
hard-codes mdns_hostname_set("MyHost"), silently overwriting whatever the
firmware set. Symptom after SSDP removal: espsomfyrts.local stopped
resolving once HAP started. Workaround in app_main re-asserts
settings.hostname and registers _http._tcp via the IDF mdns API (Arduino
MDNS.addService is wiped during hap_init). A clean SDK-fork fix is
backlogged.

clang-tidy cleanups in ControllerNetwork.cpp:
- bugprone-reserved-identifier: rename underscore-prefixed file statics
  (_lastHeapEmit, _apScanning, _lastHeap, _lastMaxHeap).
- bugprone-signed-char-misuse: NOLINT on the two RSSI int casts;
  WiFi.RSSI() is intentionally signed dBm.
- bugprone-branch-clone: simplify connected() — collapse the two
  return-false branches, drop unreachable trailing return.

Misc: EthernetSettings::usesPin replaced if/else chain with a CLKMode→pin
lookup table; ConfigFile::writeString/writeVarString slen typed as size_t.

+++ main/ConfigFile.cpp
@@ -178,7 +178,7 @@ bool ConfigFile::readVarString(char *buff, size_t len) {
 }
 bool ConfigFile::writeString(const char *val, size_t len, const char tok) {
   if(!this->isOpen()) return false;
-  int slen = strlen(val);
+  size_t slen = strlen(val);
   if(slen > 0)
     if(this->file.write((uint8_t *)val, slen) != slen) return false;
   // Now we need to pad the end of the string so that it is of a fixed length.
@@ -192,7 +192,7 @@ bool ConfigFile::writeString(const char *val, size_t len, const char tok) {
 }
 bool ConfigFile::writeVarString(const char *val, const char tok) {
   if(!this->isOpen()) return false;
-  int slen = strlen(val);
+  size_t slen = strlen(val);
   this->writeChar(CFG_TOK_QUOTE);
   if(slen > 0) if(this->file.write((uint8_t *)val, slen) != slen) return false;
   this->writeChar(CFG_TOK_QUOTE);
